Avoid `.html` suffix in URLs
Some servers e.g. GitHub, do translate un-suffixed paths, but most don't, including `simple-http-server`, which I use fro development. If we could rely on that everywhere, we could just change the link translation in the Pandoc step, without affecting the structure of `dist`. But this approach, of using `index.html` files, is universal, and will probably work better anyway when we want per-page directories with images etc.

Note that `htmlInToOut` is essntially the inverse of `htmlOutToIn`, whereas `htmlInToOut'` is just for translator link targets. Also, `joinPath . drop 1 $ splitPath` is just an easy way of dropping `outDir`, so that we can now retain full paths during the processing steps in both directions.