Use AVIF instead of JPEG for profile picture
The served file has gone from 67KB to 53KB and yet is far clearer.

Using AVIF for the checked in source image is not strictly necessary, but the file is over 10% smaller and has no noticeable loss of fidelity, and is preferable for simplicity and presumably build performance. It was created with `magick me.jpg -quality 100 me.avif`.