Revert "Disable bouncy scrolling"
This reverts commit b2013cb557398e3c4b9bbd6f6081e7d33c782077.

For some reason, it disables my ability to use two fingers on a laptop touchpad to go back a page. I don't care about the scrolling behaviour enough to investigate this.

There may also be some concern around performance of the universal selector, though I believe this is a myth with modern browsers.